requirements.txt
firebase_admin==6.6.0
Flask==3.1.0
Flask_Cors==4.0.0
flask_sqlalchemy==3.1.1
fuzzywuzzy==0.18.0
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
